fix:YoloModel removal
