+++ deepface/commons/weight_utils.py
@@ -127,7 +127,7 @@ def download_all_models_in_one_shot() -> None:
         MODEL_URL as SSD_MODEL,
         WEIGHTS_URL as SSD_WEIGHTS,
     )
-    from deepface.models.YoloModel import (
+    from deepface.models.face_detection.Yolo import (
         WEIGHT_URLS as YOLO_WEIGHTS,
         WEIGHT_NAMES as YOLO_WEIGHT_NAMES,
         YoloModel
